techDebt(buRefactorTraining): Move all training entity validation into validateTraining
